Use custom workspace for running action in
There's been a few issues with Jekyll-Diff-Action
using the default `/github/workspace` directory
to run in. That's mounted in read only mode, so
any command that tried modify it failed (see PR #1
and issue #2).
To get away from all this, this PR adds an initial
step to the diff script to create our own workspace
with the appropriate permissions.